feat: add credit review transaction endpoint and process unpaid review transactions
This commit adds two main functions - one for getting unpaid reviews from the db so as to have more visibility and the other for manually processing credit transactions for unpaid reviews.
unpaid reviews are reviews that have a mergedAt field which was not processed somehow by the webhook controller. We might consider creating a cron job for processing these reviews but that should be discussed carefully considering the caveats.