refactor: pending reviews to apply different pending limits
This commit checks if a user has successfully reviewed fewer than 3 transcripts.
If true, we allow the user to claim only 1 transcript and return an error if user has already has a pending review.
If the user has successfully reviewed 3 or more transcripts, we allow to have 6 pending reviews.
If the user has 6 or more pending reviews, we return an error.