fix: offset calculation in getAllTransactions (#299)
* feat: add evaluator role (#293)

feat: add evaluator role (#292)

* feat: Add evaluator role and permissions to user model and middleware

The code changes include adding the "evaluator" role and permissions to the user model and updating the middleware to handle the new role. This allows for better access control and authorization for evaluators.

* chore: clean up user section of swagger docs

* refactor: authorize middleware function to be dynamic with user permissions

* feat: Update username parameter in findByPublicProfile endpoint

The code changes is necessary to follow RESTful OpenAPI convention where GET request shouldn't have a request body but can have query parameters

* fix: offset calculation in getAllTransactions (#298)

fix: offset calculation in getAllTransactions (#297)

fix: fix offset calculation in getAllTransactions

The offset calculation in the getAllTransactions function was incorrect, resulting in incorrect pagination of transactions from the db. This commit fixes the calculation by subtracting 1 from the page number before multiplying it by the limit.